move logic to multi-region-site in ts-scripts
